mv metric config utils
